+++ src/navigation/TopNavBar.tsx
@@ -1,0 +1,86 @@
+import { useEffect } from "react";
+import Container from "react-bootstrap/Container";
+import Nav from "react-bootstrap/Nav";
+import Navbar from "react-bootstrap/Navbar";
+import NavDropdown from "react-bootstrap/NavDropdown";
+import { Outlet, useLocation } from "react-router-dom";
+import { Button } from "react-bootstrap";
+import Logo from "../assets/logo/alog-logo.png";
+
+const LoggedInNav = () => {
+    return (
+        <>
+            <Navbar.Collapse id="basic-navbar-nav">
+                <Nav className="me-auto">
+                    <NavDropdown title="내 작업" id="basic-nav-dropdown">
+                        <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+                        <NavDropdown.Divider />
+                        <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+                    </NavDropdown>
+                    <NavDropdown title="프로젝트" id="basic-nav-dropdown">
+                        <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+                        <NavDropdown.Divider />
+                        <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+                    </NavDropdown>
+                    <NavDropdown title="팀" id="basic-nav-dropdown">
+                        <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+                        <NavDropdown.Divider />
+                        <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+                    </NavDropdown>
+                    <Button>만들기</Button>
+                </Nav>
+            </Navbar.Collapse>
+            <Navbar.Collapse className="justify-content-end">
+                <Button>Log In</Button>
+                <Button>Sign Up</Button>
+            </Navbar.Collapse>
+        </>
+    );
+};
+
+const LoggedOutNav = () => {
+    return (
+        <>
+            <Navbar.Collapse id="basic-navbar-nav">
+                <Nav className="me-auto">
+                    <NavDropdown title="Company" id="basic-nav-dropdown">
+                        <NavDropdown.Item href="#action/3.1">Action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.2">Another action</NavDropdown.Item>
+                        <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
+                        <NavDropdown.Divider />
+                        <NavDropdown.Item href="#action/3.4">Separated link</NavDropdown.Item>
+                    </NavDropdown>
+                </Nav>
+            </Navbar.Collapse>
+            <Navbar.Collapse className="justify-content-end">
+                <Navbar.Text>알람</Navbar.Text>
+                <Navbar.Text>설정</Navbar.Text>
+                <Navbar.Text>프로필</Navbar.Text>
+            </Navbar.Collapse>
+        </>
+    );
+};
+export const TopNavBar = () => {
+    const location = useLocation();
+    return (
+        <div className="TopNavBar">
+            <Navbar expand="lg" className="bg-body-tertiary">
+                <Container>
+                    <Navbar.Brand href="/">
+                        <img src={Logo} width="30" height="30" className="d-inline-block align-top" alt="" />
+                        A-Log
+                    </Navbar.Brand>
+                    <Navbar.Toggle aria-controls="basic-navbar-nav" />
+                    {location.pathname === "/" ? <LoggedOutNav /> : <LoggedInNav />}
+                </Container>
+            </Navbar>
+            <Outlet />
+        </div>
+    );
+};
